+++ cppsim/lib/utils/uid.cpp
@@ -1,7 +1,7 @@
 #include "utils/uid.h"
 
-static int64_t id = 0;
+static int32_t id = 0;
 
-int64_t UID::next_id(){
+int32_t UID::next_id(){
     return id++;
 }
